dom/tests/mochitest/localstorage/test_bug746272-2.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>incomplete UTF-16 test</title>
<meta http-equiv="Content-type" content="text/html; charset=UTF-8" />

<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

function startTest()
{
  var test2 = localStorage.getItem("test2");
  is(test2, "value2", "expected item");

  localStorage.removeItem("test2");
  is(localStorage.length, 2, "expected number of items");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body onload="startTest();">

</body>
</html>
